Moving visualisers into a single file
